ability to change the roles and protect to delete feature

+++ convex/files.ts
@@ -2,7 +2,7 @@ import { ConvexError, v } from "convex/values";
 import { MutationCtx, QueryCtx, mutation, query } from "./_generated/server";
 import { getUser } from "./users";
 import { fileTypes } from "./schema";
-import { Id } from "./_generated/dataModel";
+import { Doc, Id } from "./_generated/dataModel";
 
 export const generateUploadUrl = mutation(async (ctx) => {
   const identity = await ctx.auth.getUserIdentity();
@@ -36,7 +36,8 @@ async function hasAccessToOrg(
   }
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -118,6 +119,16 @@ export const getFiles = query({
   },
 });
 
+// function assertCanDeleteFile(user: Doc<"users">, file: Doc<"files">) {
+//   const canDelete =
+//     file.userId === user._id ||
+//     user.orgIds.find((org) => org.orgId === file.orgId)?.role === "admin";
+
+//   if (!canDelete) {
+//     throw new ConvexError("you have no acces to delete this file");
+//   }
+// }
+
 export const deleteFile = mutation({
   args: { fileId: v.id("files") },
   async handler(ctx, args) {
@@ -126,6 +137,12 @@ export const deleteFile = mutation({
     if (!access) {
       throw new ConvexError("no access to file");
     }
+     const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === "admin";
+
+     if(!isAdmin) {
+      throw new ConvexError("you have no admin access to delete");
+     }
+    // assertCanDeleteFile(access.user, access.file);
 
     await ctx.db.delete(args.fileId);
   },
